move search form component to search reslut page
