+++ components/boxes/SearchBox/index.js
@@ -1,7 +1,5 @@
-import SearchForm from './SearchForm';
 import SearchFormTopMenu from './SearchFormTopMenu';
 
 export {
-  SearchForm,
   SearchFormTopMenu
 };
